server/requirements.txt
Flask==2.3.2
Flask-CORS==4.0.0
python-dotenv==1.0.0
cloudinary==1.36.0
Pillow==10.4.0
numpy==1.26.4
opencv-python==4.9.0.80
requests==2.31.0
gunicorn==21.2.0